move to custom_predicates, removed #has_descendante
